requirements.txt
Django==6.0
argon2-cffi==25.1.0
django-cleanup==9.0.0
django-environ==0.12.0
psycopg[binary]==3.3.2
django-storages[s3]==1.14.6
djangorestframework==3.16.1
